♻️ refactor(common): deduplicate table formatting CLI args
The table formatting options (table-format, sub-table-spacing,
separate-root-table, expand-tables, collapse-tables,
skip-wrap-for-keys) and their helper functions (_spacing_argument,
_list_argument) were identically defined in both pyproject-fmt and
tox-toml-fmt. Moved them to toml-fmt-common where they belong
alongside column-width and indent.
